Fix
Issue: "Warning: Attempt to present MMAlertVC on Reminder whose view is not in the window hierarchy!"

dismiss(animated: true) was being called as well as the presentAlertOnMainThread method. This caused the VC to be dismissed on which the alert should be presented.

Fix: Insetad of calling dismiss on the saveButtonTapped method it is now being called inside the saveReminder methods after the guard statements. This way it will never be called if there is still a error that needs to be handled by the user.