SearchResults didSelect
Issue: Searchresults without a streetnumber in the address were not selectable.

Fix: Instead of using guard statements optional binding is used.

The title for a location is no longer repeated inside the subtitle. The logic for creating a subtitle has been  moved  to a helper to clean up the searchResults delegate method.