Fix + Refactor
Issue: Single trigger and/or deleted notifications did not have the pending notification requests removed from the notifications center.

Fix: removePendingNotificationRequests should now deal with both cases. It gets called in the delete contextualAction as well as in the setInactiveStatus method inside the NotifcationManager.

Refactored the xcdatamodeld reminder entity.

- No longer uses a reminder title.
- Now uses a UUID as an identfier

As a result the UI for the ReminderVC, reminderListVC and the annotationVC has been altered, removing all use cases.